requirements.txt
flask==2.3.2
transformers==4.40.1
torch==2.2.2
tensorflow==2.15.0
numpy==1.24.3
opencv-python==4.8.0.76
gunicorn==21.2.0
pillow==10.3.0